Separating VarmintWidgetAnimation from a strait animation
